Project/farmtech/requirements.txt
# requirements.txt
flask
psutil
Werkzeug
gunicorn
flask_session==0.5.0
psycopg2-binary
redis
requests
pandas
scikit-learn